confidential-healthcare-scone-confinf-onnx/requirements.txt
flask
redis
flask_restful
pillow
numpy
cffi